requirements.txt
tensorflow>=2.13.0
numpy>=1.24.0
pillow>=10.0.0
flask>=3.0.0
flask-cors>=4.0.0
matplotlib>=3.8.0
scikit-learn>=1.3.0
opencv-python>=4.8.0
